fix team comments
